refactor(site): generate-llms 核心逻辑抽为纯 JS 方法与 vite 解耦
- common-docs/plugins/generate-llms/index.ts 改为纯 async 方法 generateLlmsDocs
  - 移除 vite 依赖（ResolvedConfig），仅依赖 fs 与 gray-matter
  - 显式接收 componentsRoot / outputDir 路径，产物落盘并返回文档列表
- 各站点 vite 插件（miniprogram / miniprogram-chat / uniapp / uniapp-chat）
  薄封装该纯方法，在 closeBundle 钩子里调用，与 changelog2json 结构一致
- 更新 types.ts 与 README
